+++ src/server/TCMine.Server.Infrastructure.Sqlite/Migrations/20260822184522_UnboundUpstreamSnapshot.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace TCMine.Server.Infrastructure.Sqlite.Migrations
+{
+    /// <inheritdoc />
+    public partial class UnboundUpstreamSnapshot : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+
+        }
+    }
+}